Added AchievementStepView

+++ ReminderXamarin/ReminderXamarin/ReminderXamarin/ViewModels/AchievementViewModel.cs
@@ -13,6 +13,6 @@ namespace ReminderXamarin.ViewModels
         public int Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public int GeneralTimeSpent { get; set; }
+        public double GeneralTimeSpent { get; set; }
     }
 }
