+++ ReminderXamarin/Rm.Data/Data/Entities/AchievementStep.cs
@@ -15,7 +15,7 @@ namespace Rm.Data.Data.Entities
         public int Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public int TimeSpent { get; set; }
+        public double TimeSpent { get; set; }
         public DateTime AchievedDate { get; set; }
 
         [ForeignKey(typeof(AchievementModel))]
